+++ hpatchz.c
@@ -78,7 +78,14 @@
 #endif
 #if (_IS_NEED_DEFAULT_CompressPlugin)
 //===== select needs decompress plugins or change to your plugin=====
-#   define _CompressPlugin_zlib
+#   ifndef _IS_NEED_decompressor_ldef_replace_zlib
+#       define _IS_NEED_decompressor_ldef_replace_zlib  0 
+#   endif
+#   if (_IS_NEED_decompressor_ldef_replace_zlib)
+#       define _CompressPlugin_ldef  //optimized zlib&ldef's deflate code decompress speed by libdeflate
+#   else
+#       define _CompressPlugin_zlib
+#   endif
 #   define _CompressPlugin_bz2
 #   define _CompressPlugin_lzma
 #   define _CompressPlugin_lzma2
@@ -93,6 +100,14 @@
 #   define _CompressPlugin_brotli
 #   define _CompressPlugin_lzham
 #   define _CompressPlugin_tuz
+#endif
+#ifdef _CompressPlugin_ldef
+#   ifndef _CompressPlugin_ldef_is_use_zlib
+#       define _CompressPlugin_ldef_is_use_zlib         1 //now ldef need zlib decompressor for any all of deflate code
+#   endif
+#   if (_IS_NEED_decompressor_ldef_replace_zlib&&(defined(_CompressPlugin_zlib)))
+#       undef _CompressPlugin_zlib
+#   endif
 #endif
 
 #if (_IS_NEED_BSDIFF)
@@ -781,8 +796,12 @@ int hpatch_cmd_line(int argc, const char * argv[]){
 #define _try_rt_dec(dec) { if (dec.is_can_open(compressType)) return &dec; }
 
 static const hpatch_TDecompress* __find_decompressPlugin(const char* compressType){
-#ifdef  _CompressPlugin_zlib
+#if ((defined(_CompressPlugin_ldef))&&_IS_NEED_decompressor_ldef_replace_zlib)
+    _try_rt_dec(ldefDecompressPlugin);
+#else
+#  ifdef  _CompressPlugin_zlib
     _try_rt_dec(zlibDecompressPlugin);
+#  endif
 #endif
 #ifdef  _CompressPlugin_bz2
     _try_rt_dec(bz2DecompressPlugin);